Corrected no ETH required clause
A lot of voters were confused about not being able to vote because there was no ETH in their wallet, they had been pointing to the clause which states no ETH is required for voting. But we have to account for gas fee, and a small amount of ETH is required with the current voting system to be able to send the transaction that contains the vote.